requirements.txt
openai
flask
flask[async]
flask_sqlalchemy
sqlalchemy
semantic-kernel
pyodbc
fastapi
azure-keyvault-secrets
azure-identity
azure-monitor-opentelemetry
Flask-Session2
werkzeug>=2
requests>=2,<3
identity>=0.5.1,<0.6
redis
redis-entraid>=0.2.0